Removed Spring-boot-starter-cache dependencies

+++ dolphinscheduler-datasource-plugin/dolphinscheduler-datasource-api/src/main/java/org/apache/dolphinscheduler/plugin/datasource/api/datasource/BaseDataSourceParamDTO.java
@@ -117,7 +117,7 @@ public abstract class BaseDataSourceParamDTO implements Serializable {
         }
 
         this.host = String.join(Constants.COMMA, hosts);
-        this.port = Integer.parseInt(portString);
+        this.port = (Integer) Integer.parseInt(portString);
     }
 
     public String getDatabase() {
